Correct description of implementation tagging
I think we initially missed a detail of the spec. As per
[TRC-IMPL.1::PREFIX.1], an implementation unit is to be tagged with
it's own pipe-designated tag that adds a suffix.

This is important for two reasons:

1. It makes tagging implementation units trivial, in that we don't need
to introduce new syntax (such as the elsewhere proposed `## Implements`
header).
2. It provides a unique tag for each implementation unit, which is
critical for when we decompose a spec into many smaller implementation
aspects, and for allowing unique reference from verification units, or
other implementation units (e.g., when using a reference implementationl
as a spec).